Simple user registration prompt
I was experimenting with user input and type conversion to create a simple user registration form. The form asks the user to enter their first and last name as strings. It then asks the user to enter her birth year -- and calculates her age by subtracting the birthyear from the current year, which has been stored in a variable.