Iteration 2 of Zen of Python poem script
This is a refactoring of my first attempt at printing the Zen of Python poem. Before, I stored each line of the poem in a variable -- and called that variable in separate print commands. This time, I am storing it in a list. And using Python's syntax to print items 0 to 18 in that list.