Iteration 3 of Zen of Python poem script
This is my second refactoring (or third iteration) of the Zen of Python poem script. Rather than storing the poem inside a function (which essentially hardcodes it and prevents any user from changing it), I am storing it in a .txt file. I have then created a function that reads the file -- and prints its contents. I like how elegant this solution is given my (still) limited knowledge in Python.