Guess and check the cube root of a value
This script uses the Guess and Check pattern in Python to calculate the cube root of a value. At this stage, the value is hardcoded as an integer in the 'cube' variable. At a later stage, I will code a next version of the script that prompts the user to input a value -- and tries to guess-and-check its cube root.